+++ R/loo.R
@@ -6,13 +6,13 @@
 #'
 #' @details The parameter \code{exclusion} defines what is left out.
 #' The value "interaction" means that a single interaction is removed.
-#' In the case of a homogenous model, this can be interpreted as the
+#' In the case of a homogeneous model, this can be interpreted as the
 #' removal of the interaction between two edges. The values "row" and
 #' "column" mean that all interactions for a row edge resp. a column
 #' edge are removed. The value "both" removes all interactions for
 #' a row and a column edge.
 #'
-#' In the case of a homogenous model, "row" and "column" don't make sense
+#' In the case of a homogeneous model, "row" and "column" don't make sense
 #' and will be replaced by "both" with a warning. This can be interpreted
 #' as removing vertices, i.e. all interactions between one edge and
 #' all other edges. For more information, see Stock et al. (2018).
@@ -45,7 +45,7 @@
 #' @rdname loo
 #' @export
 setMethod("loo",
-          "tskrrHeterogenous",
+          "tskrrHeterogeneous",
           function(x,
                    exclusion = c("interaction","row","column","both"),
                    replaceby0 = FALSE){
@@ -88,7 +88,7 @@ setMethod("loo",
 #' @rdname loo
 #' @export
 setMethod("loo",
-          "tskrrHomogenous",
+          "tskrrHomogeneous",
           function(x,
                    exclusion = c("interaction","both"),
                    replaceby0 = FALSE){
@@ -109,7 +109,7 @@ setMethod("loo",
             Hk <- hat(x)
 
             if(exclusion == "interaction"){
-              loofun <- .getloo_homogenous("interaction",
+              loofun <- .getloo_homogeneous("interaction",
                                            replaceby0,
                                            symm)
               out <- loofun(x@y, Hk, x@pred)